files/requirements.txt
serial
pyowm
configparser
paho-mqtt
